pyproject.toml
[tool.poetry]
name = "hackathon-be"
version = "0.1.0"
description = ""
authors = ["RG1ee <niwows@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"
ruff = "^0.0.265"
django-stubs = "^4.2.0"
djangorestframework-stubs = "^3.14.0"
mypy = "^1.2.0"
ipython = "^8.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.django-stubs]
django_settings_module = "src.hackathon.settings"
ignore_missing_model_attributes = true

[tool.poetry.scripts]
hackathon = "src.manage:main"
